+++ Peer.Domain/Commands/ConfigEditError.cs
@@ -1,0 +1,8 @@
+namespace Peer.Domain.Commands;
+
+public enum ConfigEditError
+{
+    Fire,
+    UnsupportedOs,
+    ProcessFailedToOpen,
+}
