+++ Peer.Domain/IFileOperations.cs
@@ -1,0 +1,9 @@
+using System.IO;
+
+namespace Peer.Domain;
+
+public interface IFileOperations
+{
+    bool Exists(string path);
+    FileStream Create(string path);
+}
